Bye, bye, libtldl.
Really, get it installed in your system or forget about it. It is mostly harmful to have the libltdl as a copy in mpg123.
We seem to get rpms with libltdl in it where there is a libtool rpm for that... needless nuisance. Get your libltdl and then mpg123 -- it's Free!



git-svn-id: svn://scm.orgis.org/mpg123/trunk@1980 35dc7657-300d-0410-a2e5-dc2837fedb53