More verbosity on HTTP action (print all server responses with -vvv) and bail out early when relocation headers don't provide a new location.
This is related to issue 2904736: The server in that case does not give

Location: http://some/place

but

location: http://some/place

All other headers are uppercase as it should be, so, rather than making mpg123 fuzzy match the broken header, I decided to make it quickly error out with a better assessment of the situation.
The server should be fixed to put out correct headers.



git-svn-id: svn://scm.orgis.org/mpg123/trunk@2450 35dc7657-300d-0410-a2e5-dc2837fedb53